+++ agrifoodpy/food/__init__.py
@@ -1,1 +1,1 @@
-from . import food_supply
+from . import food, model
