--- agrifoodpy/food/tests/test_interventions.py
@@ -1,7 +1,0 @@
-import numpy as np
-import xarray as xr
-
-import warnings
-
-def test_scale_consumption():
-    pass
