+++ agrifoodpy/food/tests/test_model.py
@@ -1,0 +1,7 @@
+import numpy as np
+import xarray as xr
+
+import warnings
+
+def test_scale_consumption():
+    pass
