+++ agrifoodpy/land/land.py
@@ -3,33 +3,7 @@
 
 import numpy as np
 import xarray as xr
-import os
 import matplotlib.pyplot as plt
-from itertools import product
-
-data_dir = os.path.join(os.path.dirname(__file__), 'data/' )
-
-available = [
-    "ALC_ALC_1000",
-    "ALC_ALC_2000",
-    "ALC_ALC_5000",
-    "CEH_LC_1000",
-    "CEH_LCPCP_1000",
-    "CEH_LCPCP_2000",
-    "CEH_LCPCP_5000",
-]
-
-def __getattr__(name):
-    if name not in available:
-        raise AttributeError(f"{name!r} does not match a dataset and resoliution in {__name__!r}.")
-
-    origin, dataset, resolution = name.split('_')
-    _data_file = f'{data_dir}{origin}/{dataset}_{resolution}.nc'
-    
-    data = xr.open_dataset(_data_file)
-
-    return data
-
 
 @xr.register_dataarray_accessor("land")
 class LandDataArray:
@@ -49,7 +23,21 @@ class LandDataArray:
             raise AttributeError("Land array must have 'x' and 'y' dimensions")
 
     def plot(self, ax=None, **kwargs):
-        """Plot 2D map using imshow, without interpolation and without
+        """Plot a LandDataArray
+        
+        Generates a plot of a LandDataArray using matplotlib imshow, without
+        interpolation and setting the origin low to align north at the top.
+
+        Parameters
+        ----------
+        ax : matplotlib.pyplot.Artist
+            Axes on which to draw the plot
+        **kwargs : dict
+            Style options to be passed to the imshow function.
+        
+        Returns
+        -------
+        ax : matplotlib axes instance
         """
 
         map = self._obj
@@ -64,9 +52,29 @@ class LandDataArray:
         ax.imshow(map, interpolation="none", origin="lower",
                   extent=[xmin, xmax, ymin, ymax])
         
+        return ax
+        
     def area_by_type(self, values = None, dim = None):
-        """Returns the total number of pixels where the value equals the input
-        list of values.        
+        """Area per map category in a LandDataArray
+        
+        Returns a DataArray with the total number of pixels for each category or
+        category subset of the LandDataArray. 
+
+        Parameters
+        ----------
+        values : int, array
+            List of category types to return the total area for. If not set, the
+            function returns areas for all values found on the map, excluding
+            nan values.
+        dim : string
+            Name to assign to the categories coordinate. If not set, the input
+            DataArray name is used instead.
+        
+        Returns
+        -------
+        xarray.DataArray
+            Array with the corresponding areas overlaps for each category
+            combination.
         """
 
         map = self._obj
@@ -77,11 +85,12 @@ class LandDataArray:
 
         if values is None:
             values = np.unique(map)
+        else:
+            values = np.array(values)
 
         # Prevent nan values from being counted
         nan_indices = np.isnan(values)
         values = values[~nan_indices]
-
         area = [ones.where(map==value).sum() for value in values]
         
         area_arr = xr.DataArray(area, dims=dim, coords={dim:values})
@@ -89,11 +98,40 @@ class LandDataArray:
         return area_arr
 
     def area_overlap(self, map_right, values_left = None, values_right = None, dim_left=None, dim_right=None):
+        """Area overlap of selected categories between two maps
+        
+        Returns a DataArray with the total number of pixels for each combination
+        of categories from the left and right map selected categories. Casa
 
+        Parameters
+        ----------
+        map_right : xarray.DataArray
+            LandDataArray style DataArray to compare overlapping areas with
+        values_left
+            List of category types from the left map to return the total area
+            overlaps for.
+            If not set, all category types are used, except nan values.
+        values_right : int, array
+            List of category types from the right map to return the total area
+            overlaps for.
+            If not set, all category types are used, except nan values.
+        dim_left : string
+            Names to assign to the category coordinates on the output DataArray.
+            If not set, the input DataArray name is used instead.
+        dim_right : string
+            Names to assign to the category coordinates on the output DataArray.
+            If not set, the input DataArray name is used instead.
+        
+        Returns
+        -------
+        area_arr : xarray.DataArray
+            Array with the corresponding areas for each category type
+
+        """
         map_left = self._obj
 
         # Check that both maps have the same dimensions and coordinates. if not,
-        # this raises a ValueError
+        # this raises a ValueError (alternatively, align the maps and use )
         xr.align(map_left, map_right, join='exact')
 
         if dim_left is None:
@@ -104,9 +142,13 @@ class LandDataArray:
 
         if values_left is None:
             values_left = np.unique(map_left)
+        else:
+            values_left = np.array(values_left)
 
         if values_right is None:
             values_right = np.unique(map_right)
+        else:
+            values_right = np.array(values_right)
 
         # Prevent nan values from being counted
         nan_indices_left = np.isnan(values_left)
@@ -121,20 +163,37 @@ class LandDataArray:
         area_arr = xr.DataArray(area, dims=[dim_left, dim_right], coords={dim_left:values_left, dim_right:values_right})
 
         return area_arr
-
-@xr.register_dataset_accessor("land_ds")
-class LandDataset:
-    def __init__(self, xarray_obj):
-        self._obj = xarray_obj
-        self._validate(xarray_obj)
-
-    @staticmethod
-    def _validate(obj):
-        """Validate land_ds xarray, checking it is a DataArray and has the minimum set
-        of arrays
-        """
-        if not isinstance(obj, xr.Dataset):
-            raise TypeError("Land array must be an xarray.Dataset")
     
-        if "x" not in obj.dims or "y" not in obj.dims:
-            raise AttributeError("Land array must have 'x' and 'y' dimensions")
+    def category_match(self, map_right, values_left=None, values_right=None,
+                       join="left"):
+        """Returns a land Dataarray with values where a selected overlap occurs
+        between categories from two maps.
+        Values are retained from the left map.
+
+        Parameters
+        ----------
+        map_right : xarray.DataArray
+            LandDataArray style DataArray to compare overlapping areas with
+        values_left : int, array
+            List of category types from the left map to match.
+            If not set, all category types are used, except nan values.
+        values_right : int, array
+            List of category types from the right map to match.
+            If not set, all category types are used, except nan values.
+                
+        Returns
+        -------
+        category_match : xarray.DataArray
+            Land DataArray with values from the left map where overlap occurs.
+            All other positions are assign a nan value.
+        """
+        map_left = self._obj
+
+        # Align maps to the left so they have the same dimensions
+        map_left, map_right = xr.align(map_left, map_right, join=join)
+
+        shape = map_left.shape
+
+        category_match = map_left.where(np.in1d(map_left, values_left).reshape(shape)).where(np.in1d(map_right, values_right).reshape(shape))
+
+        return category_match
