+++ agrifoodpy/land/tests/test_model.py
@@ -1,0 +1,6 @@
+import numpy as np
+import xarray as xr
+
+def test_carbon_sequestration():
+    pass
+
